Time: 0 ms (100.00%) | Memory: 7.8 MB (26.97%) - LeetSync

+++ 2021-remove-all-occurrences-of-a-substring/remove-all-occurrences-of-a-substring.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     string removeOccurrences(string s, string part) {
-        size_t pos = s.find(part);
+        int pos = s.find(part);
 
         while (pos != string::npos) {
             s.erase(pos, part.size());
